Undo modules rename
