+++ quantumflow/__init__.py
@@ -55,7 +55,7 @@ from .decompositions import *  # noqa: F401, F403
 from .gates import *  # noqa: F401, F403
 from .gradients import *  # noqa: F401, F403
 from .info import *  # noqa: F401, F403
-from .multigates import *  # noqa: F401, F403
+from .modules import *  # noqa: F401, F403
 from .ops import *  # noqa: F401, F403
 from .paulialgebra import *  # noqa: F401, F403
 from .qubits import *  # noqa: F401, F403
